+++ docs/MunicipalityRegistry.Structurizr/Program.cs
@@ -61,7 +61,8 @@ namespace MunicipalityRegistry.Structurizr
         {
             var configuration = new ConfigurationBuilder()
                 .SetBasePath(Directory.GetCurrentDirectory())
-                .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
+                .AddJsonFile("appsettings.json", optional: false, reloadOnChange: false)
+                .AddJsonFile($"appsettings.{Environment.MachineName.ToLowerInvariant()}.json", optional: true, reloadOnChange: false)
                 .Build();
 
             _workspaceId = long.Parse(configuration["Structurizr:WorkspaceId"]);