+++ src/MunicipalityRegistry.Api.CrabImport/Infrastructure/EmptyController.cs
@@ -19,7 +19,7 @@ namespace MunicipalityRegistry.Api.CrabImport.Infrastructure
         {
             return Request.Headers[HeaderNames.Accept].ToString().Contains("text/html")
                 ? (IActionResult)new RedirectResult("/docs")
-                : new OkObjectResult($"Welcome to the VBR Municipality Import Api v{Assembly.GetEntryAssembly().GetName().Version}.");
+                : new OkObjectResult($"Welcome to the Basisregisters Vlaanderen Municipality Api v{Assembly.GetEntryAssembly().GetName().Version}.");
         }
     }
 }